requirements.txt
aiofiles==24.1.0
aiogoogle==5.13.2
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
attrs==24.3.0
bcrypt==4.2.1
black==24.10.0
cachetools==5.5.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-filter==2.0.1
fastapi-users==14.0.0
fastapi-users-db-sqlalchemy==6.0.1
flake8==7.1.1
frozenlist==1.5.0
google-auth==2.37.0
greenlet==3.1.1
h11==0.16.0
httptools==0.6.4
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.3
makefun==1.15.6
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.1
pwdlib==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.27.2
pyflakes==3.2.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqladmin==0.16.1
SQLAlchemy==2.0.36
starlette==0.41.3
tonyg-rfc3339==0.1
typing-inspection==0.4.0
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.3
websockets==14.1
WTForms==3.1.2
yarl==1.18.3
